+++ project/bodquest_2023/lib/presentation/component/main/main_widget.dart
@@ -3,9 +3,9 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../notifier/index_notifier.dart';
 import '../../page/home_page.dart';
-import '../../page/training_page.dart';
+import '../../page/training/training_page.dart';
 import '../../page/user_page.dart';
-import '../../page/weight_page.dart';
+import '../../page/weight/weight_page.dart';
 import '../../page/health_sample_page.dart';
 import '../../page/sample_counter_page.dart';
