+++ project/bodquest_2023/lib/presentation/page/training/training_page.dart
@@ -1,16 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../component/component_types.dart';
-import '../component/control/number_textfield.dart';
-import '../component/training/training_kind_dropdown.dart';
-import '../component/training/training_list_item.dart';
-import '../notifier/datetime_notifier.dart';
-import '../notifier/training/training_list_provider.dart';
-import '../notifier/text_notifier.dart';
-import '../notifier/training/training_kind_notifier.dart';
-import '../notifier/user/login_user_provider.dart';
-import '../state/training/training_state.dart';
+import '../../component/component_types.dart';
+import '../../component/control/number_textfield.dart';
+import '../../component/training/training_kind_dropdown.dart';
+import '../../component/training/training_list_item.dart';
+import '../../notifier/datetime_notifier.dart';
+import '../../notifier/training/training_list_provider.dart';
+import '../../notifier/text_notifier.dart';
+import '../../notifier/training/training_kind_notifier.dart';
+import '../../notifier/user/login_user_provider.dart';
+import '../../state/training/training_state.dart';
 
 class TrainingPage extends ConsumerStatefulWidget {
   const TrainingPage({super.key});
